[#285] List functions Haddock Table
Resolves #285
