+++ CHANGELOG.md
@@ -59,6 +59,8 @@ The changelog is available [on GitHub][2].
   + Replace `mapToSnd` with `toSnd`
   + You can now use `fmapToFst` and `fmapToSnd` instead of
     `[f]map (mapToFst f)` and `[f]map (mapToSnd f)`
+* Reexport `NonEmpty` functions from `Relude.List.NonEmpty` instead of
+  `Relude.List.Reexport`.
 
 ## 0.6.0.0 — Oct 30, 2019
